fix: refactor init module
